Part 5 circular import fixes
